subscription will now only be present in chat and after opening simulator
