+++ src/containers/Chat/ChatInterface/ChatInterface.module.css
@@ -1,0 +1,106 @@
+.Chat {
+  width: 100%;
+  display: flex;
+}
+
+.ChatMessages {
+  width: 100%;
+}
+
+.ChatConversations {
+  width: 339px !important;
+  min-width: 339px !important;
+  border-left: 1px solid rgba(0, 0, 0, 0.12);
+}
+
+@media only screen and (max-width: 768px) {
+  .ChatConversations {
+    display: none !important;
+    width: 100% !important;
+  }
+  .TabContainer {
+    display: flex !important;
+    width: 100% !important;
+    justify-content: center;
+  }
+  .ToolBar {
+    padding: 24px 16px 0 16px !important;
+    width: 100% !important;
+  }
+}
+
+.NoConversations {
+  padding: 76px;
+}
+
+.IconBackground {
+  height: 24px;
+  margin-left: auto;
+  margin-right: auto;
+  width: 50%;
+}
+
+.Icon {
+  margin: 0px 8px 8px 8px;
+}
+
+.Title {
+  display: inline-block;
+  width: 100%;
+  text-align: center;
+}
+
+/* Overriding Typography styling */
+.TitleText {
+  font-size: 16px !important;
+  margin: 4px 0px 5px 0px !important;
+  color: #93a29b;
+  cursor: pointer;
+  line-height: 1 !important;
+}
+
+.SelectedTab {
+  color: #073f24 !important;
+}
+
+.ToolBar {
+  padding: 24px 16px 0 16px !important;
+  width: 338px;
+}
+.TabContainer {
+  display: -webkit-inline-box;
+  width: 345px;
+}
+
+.TabContainer > div {
+  width: 85px;
+  margin-right: 12px;
+}
+
+.TabContainer > div:last-child {
+  width: auto;
+  margin-right: 0px;
+}
+
+.Title a {
+  text-decoration: none;
+  color: inherit;
+}
+
+.LightHighLighter {
+  display: flex;
+  align-items: center;
+  background-color: #93a29b;
+  height: 1px;
+  margin: 2px -6px;
+}
+
+.DarkHighLighter {
+  display: flex;
+  align-items: center;
+  background-color: #073f24;
+  height: 4px;
+  margin: 0px -6px;
+  border-top-left-radius: 10px;
+  border-top-right-radius: 10px;
+}
